Move the shared components under modules, and clean up the common code
