--- common/css/vip-workflow-admin.css
@@ -1,5 +1,0 @@
-/* Admin menu icon size */
-#adminmenu .toplevel_page_vw-custom-status .wp-menu-image img {
-	width: 16px;
-	height: 16px;
-}
