+++ modules/custom-status/lib/components/workflow-manager.js
@@ -5,9 +5,9 @@ import { useRef, useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { plusCircle } from '@wordpress/icons';
 
+import SuccessNotice from '../../../shared/js/components/success-notice';
 import CustomStatusEditor from './custom-status-editor';
 import DraggableCustomStatus from './draggable-custom-status';
-import SuccessNotice from './success-notice';
 import WorkflowArrow, { useRefDimensions } from './workflow-arrow';
 
 export default function WorkflowManager( { customStatuses } ) {
