+++ modules/notifications/notifications.php
@@ -6,8 +6,8 @@
 namespace VIPWorkflow\Modules;
 
 use VIPWorkflow\VIP_Workflow;
-use VIPWorkflow\Common\PHP\Module;
-use function VIPWorkflow\Common\PHP\vw_draft_or_post_title;
+use VIPWorkflow\Modules\Shared\PHP\Module;
+use function VIPWorkflow\Modules\Shared\PHP\vw_draft_or_post_title;
 
 class Notifications extends Module {
