+++ modules/shared/js/components/modals/confirm-delete-modal.js
