+++ modules/shared/js/components/success-notice.js
